Upadated operations test runner
